APIs/NutriAI/requirements.txt
google-cloud-storage>=1.16.1
https://download.pytorch.org/whl/cpu/torch-1.0.1.post2-cp37-cp37m-linux_x86_64.whl 
numpy